+++ src/drivers/generic/resources/point/point_value_writer.py
@@ -1,12 +1,10 @@
 from abc import abstractmethod
 
 from flask_restful import reqparse
-from flask_restful.reqparse import request
-from mrb.validator import is_bridge
 from rubix_http.exceptions.exception import NotFoundException, BadDataException
 from rubix_http.resource import RubixResource
 
-from src.models.point.model_point import PointModel
+from src.drivers.generic.models.point import GenericPointModel
 
 
 class GenericPointValueWriterBase(RubixResource):
@@ -19,13 +17,13 @@ class GenericPointValueWriterBase(RubixResource):
 
     @classmethod
     @abstractmethod
-    def get_point(cls, **kwargs) -> PointModel:
+    def get_point(cls, **kwargs) -> GenericPointModel:
         raise NotImplementedError('Please override get_point method')
 
     @classmethod
     def patch(cls, **kwargs):
-        data = GenericPointValueWriterBase.patch_parser.parse_args()
-        point: PointModel = cls.get_point(**kwargs)
+        data = cls.patch_parser.parse_args()
+        point: GenericPointModel = cls.get_point(**kwargs)
         if not point:
             raise NotFoundException('Point does not exist')
         if not point.writable:
@@ -34,18 +32,18 @@ class GenericPointValueWriterBase(RubixResource):
                                  priority=data.get('priority'),
                                  value_raw=data.get('value_raw'),
                                  fault=data.get('fault'),
-                                 fault_message=data.get('fault_message'),
-                                 sync=not is_bridge(request.args))
+                                 fault_message=data.get('fault_message'))
         return {}
 
 
-class GenericUUIDPointValueWriter(GenericPointValueWriterBase):
+class GenericPointUUIDValueWriter(GenericPointValueWriterBase):
     @classmethod
-    def get_point(cls, **kwargs) -> PointModel:
-        return PointModel.find_by_uuid(kwargs.get('uuid'))
+    def get_point(cls, **kwargs) -> GenericPointModel:
+        return GenericPointModel.find_by_uuid(kwargs.get('uuid'))
 
 
-class GenericNamePointValueWriter(GenericPointValueWriterBase):
+class GenericPointNameValueWriter(GenericPointValueWriterBase):
     @classmethod
-    def get_point(cls, **kwargs) -> PointModel:
-        return PointModel.find_by_name(kwargs.get('network_name'), kwargs.get('device_name'), kwargs.get('point_name'))
+    def get_point(cls, **kwargs) -> GenericPointModel:
+        return GenericPointModel.find_by_name(kwargs.get('network_name'), kwargs.get('device_name'),
+                                              kwargs.get('point_name'))