+++ src/routes.py
@@ -10,8 +10,8 @@ from src.drivers.generic.resources.network.network_singular import GenericNetwor
 from src.drivers.generic.resources.point.point_plural import GenericPointPlural
 from src.drivers.generic.resources.point.point_singular import GenericPointSingularByUUID, \
     GenericPointSingularByName
-from src.drivers.generic.resources.point.point_value_writer import GenericUUIDPointValueWriter, \
-    GenericNamePointValueWriter
+from src.drivers.generic.resources.point.point_value_writer import GenericPointUUIDValueWriter, \
+    GenericPointNameValueWriter
 from src.drivers.modbus.resources.device.device_plural import ModbusDevicePlural
 from src.drivers.modbus.resources.device.device_singular import ModbusDeviceSingularByUUID, \
     ModbusDeviceSingularByName
@@ -27,6 +27,8 @@ from src.drivers.modbus.resources.point.point_singular import ModbusPointSingula
     ModbusPointSingularByName
 from src.drivers.modbus.resources.point.point_stores import ModbusPointPluralPointStore, ModbusPointStore, \
     ModbusDevicePointPluralPointStore
+from src.drivers.modbus.resources.point.point_sync import MPToBPSync
+from src.drivers.modbus.resources.point.point_value_writer import ModbusPointUUIDValueWriter, ModbusPointNameValueWriter
 from src.resources.resource_device import DeviceResourceByUUID, DeviceResourceByName, DeviceResourceList
 from src.resources.resource_network import NetworkResourceByUUID, NetworkResourceByName, NetworkResourceList
 from src.resources.resource_point import PointResourceByUUID, PointResourceByName, PointResourceList
@@ -64,8 +66,8 @@ api_generic.add_resource(GenericPointPlural, '/points')
 api_generic.add_resource(GenericPointSingularByUUID, '/points/uuid/<string:uuid>')
 api_generic.add_resource(GenericPointSingularByName,
                          '/points/name/<string:network_name>/<string:device_name>/<string:point_name>')
-api_generic.add_resource(GenericUUIDPointValueWriter, '/points_value/uuid/<string:uuid>')
-api_generic.add_resource(GenericNamePointValueWriter,
+api_generic.add_resource(GenericPointUUIDValueWriter, '/points_value/uuid/<string:uuid>')
+api_generic.add_resource(GenericPointNameValueWriter,
                          '/points_value/name/<string:network_name>/<string:device_name>/<string:point_name>')
 
 bp_modbus = Blueprint('modbus', __name__, url_prefix='/api/modbus')
@@ -85,15 +87,22 @@ api_modbus.add_resource(ModbusPointPollNonExisting, '/poll/point')
 api_modbus.add_resource(ModbusPointPluralPointStore, '/point_stores')
 api_modbus.add_resource(ModbusPointStore, '/point_stores/<string:uuid>')
 api_modbus.add_resource(ModbusDevicePointPluralPointStore, '/<string:device_uuid>/point_stores')
+api_modbus.add_resource(ModbusPointUUIDValueWriter, '/points_value/uuid/<string:uuid>')
+api_modbus.add_resource(ModbusPointNameValueWriter,
+                        '/points_value/name/<string:network_name>/<string:device_name>/<string:point_name>')
 
 # Modbus <> Generic|BACnet points mappings
-bp_mapping_mp_gbp = Blueprint('mappings_mp_gbp', __name__, url_prefix='/api/mp_gbp/mappings')
+bp_mapping_mp_gbp = Blueprint('mappings_mp_gbp', __name__, url_prefix='/api/mappings/mp_gbp')
 api_mapping_mp_gbp = Api(bp_mapping_mp_gbp)
 api_mapping_mp_gbp.add_resource(MPGBPMappingResourceList, '')
 api_mapping_mp_gbp.add_resource(MPGBPMappingResourceByUUID, '/uuid/<string:uuid>')
 api_mapping_mp_gbp.add_resource(MPGBPMappingResourceByModbusPointUUID, '/modbus/<string:uuid>')
 api_mapping_mp_gbp.add_resource(MPGBPMappingResourceByGenericPointUUID, '/generic/<string:uuid>')
 api_mapping_mp_gbp.add_resource(MPGBPMappingResourceByBACnetPointUUID, '/bacnet/<string:uuid>')
+
+bp_sync = Blueprint('sync', __name__, url_prefix='/api/sync')
+api_sync = Api(bp_sync)
+api_sync.add_resource(MPToBPSync, '/mp_to_bp')
 
 bp_system = Blueprint('system', __name__, url_prefix='/api/system')
 api_system = Api(bp_system)